+++ src/game/gameflow.c
@@ -511,8 +511,8 @@ static bool GameFlow_LoadLevelSequence(
         } else if (!strcmp(type_str, "remove_ammo")) {
             seq->type = GFS_REMOVE_AMMO;
 
-        } else if (!strcmp(type_str, "remove_medpacks")) {
-            seq->type = GFS_REMOVE_MEDPACKS;
+        } else if (!strcmp(type_str, "remove_medipacks")) {
+            seq->type = GFS_REMOVE_MEDIPACKS;
 
         } else if (!strcmp(type_str, "give_item")) {
             seq->type = GFS_GIVE_ITEM;
@@ -952,7 +952,7 @@ void GameFlow_Shutdown(void)
                     case GFS_PLAY_SYNCED_AUDIO:
                     case GFS_FIX_PYRAMID_SECRET_TRIGGER:
                     case GFS_REMOVE_AMMO:
-                    case GFS_REMOVE_MEDPACKS:
+                    case GFS_REMOVE_MEDIPACKS:
                         break;
                     }
                     seq++;
@@ -1083,7 +1083,7 @@ GameFlow_InterpretSequence(int32_t level_num, GAMEFLOW_LEVEL_TYPE level_type)
     g_GameInfo.remove_guns = false;
     g_GameInfo.remove_scions = false;
     g_GameInfo.remove_ammo = false;
-    g_GameInfo.remove_medpacks = false;
+    g_GameInfo.remove_medipacks = false;
 
     GAMEFLOW_SEQUENCE *seq = g_GameFlow.levels[level_num].sequence;
     GAMEFLOW_OPTION ret = GF_EXIT_TO_TITLE;
@@ -1279,9 +1279,9 @@ GameFlow_InterpretSequence(int32_t level_num, GAMEFLOW_LEVEL_TYPE level_type)
             }
             break;
 
-        case GFS_REMOVE_MEDPACKS:
+        case GFS_REMOVE_MEDIPACKS:
             if (level_type != GFL_SAVED) {
-                g_GameInfo.remove_medpacks = true;
+                g_GameInfo.remove_medipacks = true;
             }
             break;
 
@@ -1340,7 +1340,7 @@ GameFlow_StorySoFar(int32_t level_num, int32_t savegame_level)
         case GFS_REMOVE_SCIONS:
         case GFS_FIX_PYRAMID_SECRET_TRIGGER:
         case GFS_REMOVE_AMMO:
-        case GFS_REMOVE_MEDPACKS:
+        case GFS_REMOVE_MEDIPACKS:
             break;
 
         case GFS_START_GAME: