+++ src/global/types.h
@@ -1024,7 +1024,7 @@ typedef enum GAMEFLOW_SEQUENCE_TYPE {
     GFS_MESH_SWAP,
     GFS_FIX_PYRAMID_SECRET_TRIGGER,
     GFS_REMOVE_AMMO,
-    GFS_REMOVE_MEDPACKS,
+    GFS_REMOVE_MEDIPACKS,
 } GAMEFLOW_SEQUENCE_TYPE;
 
 typedef enum GAME_STRING_ID {
@@ -1550,7 +1550,7 @@ typedef struct GAME_INFO {
     bool remove_guns;
     bool remove_scions;
     bool remove_ammo;
-    bool remove_medpacks;
+    bool remove_medipacks;
 } GAME_INFO;
 
 typedef struct CREATURE_INFO {